Rebalance Catch2 test translation units
